Optimize and refactor responses

+++ src/main/java/de/th/koeln/archilab/fae/faeteam2service/heimbewohner/HeimbewohnerRepository.java
@@ -2,5 +2,5 @@ package de.th.koeln.archilab.fae.faeteam2service.heimbewohner;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface HeimbewohnerRepository extends CrudRepository<Heimbewohner, Integer> {
+public interface HeimbewohnerRepository extends CrudRepository<Heimbewohner, Long> {
 }
